Refactor PDF upload functionality: Moved upload logic to a new controller and updated routes to use the new controller methods.
