+++ routes/uploadRoutes.js
@@ -1,11 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const upload = require('../middleware/upload');
+const { uploadPDF, uploadPDFFile } = require('../controllers/uploadController');
 
 // POST /api/upload/image
-router.post('/image', upload.single('image'), (req, res) => {
-  if (!req.file) return res.status(400).json({ msg: 'No file uploaded' });
-  res.status(200).json({ filename: req.file.filename, path: `/uploads/${req.file.filename}` });
-});
+router.post('/image', uploadPDF.single('pdf'), uploadPDFFile);
 
 module.exports = router;
